move homepage containers to App.jsx
