+++ views/index.ejs
@@ -1,2 +1,3 @@
 <div id="app"></div>
 <script src="bundle.js"></script>
+<link rel="stylesheet" href="style.css">
